# docs/swagger.yaml
basePath: /
definitions:
...
  response.Account:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
  response.FormattedValidationError:
    properties:
      httpStatusCode:
        type: integer
      message:
        type: string
      validationErrors:
        items:
          $ref: '#/definitions/response.ValidationError'
        type: array
    type: object
...
  response.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
      path:
        type: string
      value: {}
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a implementation of transaction server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Transactions Server
  version: "1.0"
paths:
...
  /accounts/{accountId}:
    get:
      consumes:
      - application/json
      description: Find an account by ID
      parameters:
      - description: ID of desired account
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Account'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.FormattedValidationError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.FormattedValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.FormattedValidationError'
      summary: Find account
      tags:
      - Account
...
swagger: "2.0"